% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FA.R
\name{FA}
\alias{FA}
\title{Principal Component Analysis for Factor Models}
\usage{
FA(X, r)
}
\arguments{
\item{X}{Input matrix, of dimension \eqn{T\times N}. Each row is an observation with \eqn{N} features at time point \eqn{t}.}

\item{r}{A positive integer indicating the factor numbers.}
}
\value{
A list containing:
\item{Fhat}{The estimated factor matrix.}
\item{Lhat}{The estimated loading matrix.}
}
\description{
Principal Component Analysis for Factor Models
}
\examples{
\donttest{
dat = gendata()
FA(dat$Y, r = 3)
}
}
\author{
Jiaqi Hu
}
