% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{TR}
\alias{TR}
\title{Trace ratios}
\usage{
TR(Fhat, F0)
}
\arguments{
\item{Fhat}{The estimated factors.}

\item{F0}{The true factors.}
}
\value{
A numeric value of the trace ratios.
}
\description{
Trace ratios
}
\examples{
\donttest{
dat = gendata()
Y = dat$Y
F0 = dat$F0
res = REFA(dat$Y, r = 3)
Fhat = res$Fhat
TR(Fhat, F0)
}
}
\references{
Robust factor analysis with exponential squared loss. Jiaqi Hu, Tingyin Wang, Xueqin Wang. Journal of Multivariate Analysis 2026, 213 105567; doi:10.1016/j.jmva.2025.105567
}
\author{
Jiaqi Hu
}
